fixed a bug that caused the relationship field to be serialized as empty in the attributes section

+++ Src/Hypermedia.JsonApi/Converters/Int32Converter.cs
@@ -30,7 +30,7 @@ namespace Hypermedia.JsonApi.Converters
         /// <returns>The object that represents the CLR version of the given JSON value.</returns>
         public object Deserialize(Type type, JsonValue jsonValue)
         {
-            return ((JsonInteger)jsonValue).Value;
+            return (int)((JsonInteger)jsonValue).Value;
         }
     }
 }
